+++ lib/src/animated_marker.dart
@@ -6,7 +6,7 @@ import 'package:google_maps_flutter/google_maps_flutter.dart'
 class AnimatedMarker extends StatefulWidget {
   /// Wrap a [GoogleMap] with this [AnimatedMarker] widget
   ///
-  /// pass the [Set] of [Marker] to [markerPositions]
+  /// pass the [Set] of [Marker] to [animatedMarkers]
   ///
   /// then take the [animatedMarkers] from it's builder method and supply it
   /// to the [GoogleMap]'s [markers]
@@ -14,11 +14,14 @@ class AnimatedMarker extends StatefulWidget {
   /// this widget will then calculate and animate the [Marker] from
   /// it's old position to the new position automatically in the [duration]
   /// with a [curve]
-  const AnimatedMarker({
+  AnimatedMarker({
     super.key,
 
-    /// [Set] of [Marker] same as in [GoogleMap]
-    required this.markerPositions,
+    /// [Set] of [Marker]s to animate, same as in [GoogleMap]
+    required this.animatedMarkers,
+
+    /// [Set] of [Marker]s that are not animated, same as in [GoogleMap]
+    Set<Marker>? staticMarkers,
 
     /// build your [GoogleMap] in this builder with the [animatedMarkers]
     required this.builder,
@@ -28,8 +31,9 @@ class AnimatedMarker extends StatefulWidget {
 
     /// default [curve] of [Curves.ease]
     this.curve = Curves.ease,
-  });
-  final Set<Marker> markerPositions;
+  }) : staticMarkers = staticMarkers ?? <Marker>{};
+  final Set<Marker> animatedMarkers;
+  final Set<Marker> staticMarkers;
   final Widget Function(BuildContext context, Set<Marker> animatedMarkers)
       builder;
   final Duration duration;
@@ -52,10 +56,10 @@ class AnimatedMarkerState extends State<AnimatedMarker>
     super.initState();
 
     /// initialize the last marker positions as the input marker positions
-    _lastMarkerPositions = widget.markerPositions;
+    _lastMarkerPositions = widget.animatedMarkers;
 
     /// initialize the current marker positions as the input marker positions
-    _currentMarkerPositions = widget.markerPositions;
+    _currentMarkerPositions = widget.animatedMarkers;
 
     /// create an animation controller with [duration]
     _controller = AnimationController(
@@ -65,7 +69,7 @@ class AnimatedMarkerState extends State<AnimatedMarker>
 
     /// create a list of marker pairs of the same [MarkerId] according to the
     /// input [markerPositions]
-    _markerPairs = widget.markerPositions.map<Map<Marker, Marker>>((marker) {
+    _markerPairs = widget.animatedMarkers.map<Map<Marker, Marker>>((marker) {
       return <Marker, Marker>{
         marker: _lastMarkerPositions.firstWhere(
           (lastMarker) => lastMarker.markerId == marker.markerId,
@@ -92,7 +96,8 @@ class AnimatedMarkerState extends State<AnimatedMarker>
             .map(
               (animation) => animation.value,
             )
-            .toSet();
+            .toSet()
+          ..addAll(widget.staticMarkers);
       });
     });
   }
@@ -110,12 +115,12 @@ class AnimatedMarkerState extends State<AnimatedMarker>
     /// when the widget updates check if the old marker position is same as the
     /// new one, if not, update it with the latest
     /// and restart the animation controller
-    if (oldWidget.markerPositions != widget.markerPositions) {
-      _lastMarkerPositions = oldWidget.markerPositions;
+    if (oldWidget.animatedMarkers != widget.animatedMarkers) {
+      _lastMarkerPositions = oldWidget.animatedMarkers;
 
       /// update the list of marker pairs of the same [MarkerId] according to the
       /// input [markerPositions]
-      _markerPairs = widget.markerPositions.map<Map<Marker, Marker>>((marker) {
+      _markerPairs = widget.animatedMarkers.map<Map<Marker, Marker>>((marker) {
         return <Marker, Marker>{
           marker: _lastMarkerPositions.firstWhere(
             (lastMarker) => lastMarker.markerId == marker.markerId,